refactor: split chunker.rs (612→13 files) for SRP/50-line limit
Split crates/codetether-rlm/src/chunker.rs into focused submodules:
- mod.rs: RlmChunker facade + re-exports
- types.rs: ContentType, Chunk, ChunkType, ChunkOptions
- detect.rs: detect_content_type dispatcher
- detect_helpers.rs: is_code/log/conversation/document_line helpers
- hints.rs: get_processing_hints
- estimate.rs: estimate_tokens heuristic
- chunk.rs: main chunk() logic
- boundaries.rs: find_boundaries
- split.rs: split_large_chunk
- select.rs: select_chunks with token budget
- reassemble.rs: reassemble with gap markers
- compress.rs: compress convenience function
- tests.rs: unit tests
